+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [5.0.2] - 2015-12-17 ##
 - All logs now saved as ".txt".
 - Catch ConnectionClosedException and UncheckedIOException in HttpConnector.
+- Use first available port for REST server.
 - Changed license to Apache 2.0
 
 ## [5.0.1] - 2015-09-24 ##